+++ src/win32/vx_fs_win32.c
@@ -27,7 +27,7 @@ bool vx_fs_cp(const char *src, const char *dest)
     return CopyFileW(w_src, w_dest, FALSE) != 0;
 }
 
-bool vx_fs_ln(const char *src, const char *dest)
+bool vx_fs_ln(const char *src, const char *dest, bool replace)
 {
     wchar_t w_src[VX_PATH_MAX], w_dest[VX_PATH_MAX];
 
@@ -40,7 +40,7 @@ bool vx_fs_ln(const char *src, const char *dest)
     }
 
     DWORD error = GetLastError();
-    if (error == ERROR_ALREADY_EXISTS || error == ERROR_FILE_EXISTS)
+    if (replace && (error == ERROR_ALREADY_EXISTS || error == ERROR_FILE_EXISTS))
     {
         DeleteFileW(w_dest);
         return CreateHardLinkW(w_dest, w_src, NULL) != 0;
